REVERTME: [VPG]: drm/i915: Notifying audio driver on hotplug/unplug
This patch updates the register values and encoder specific information
at the time of hotplug and unplug which is required by audio driver for
registering audio. This information is then notified to the audio driver
on hotplug/unplug of the external device.

REVERTME: LP audio is not available upstream and may not be for
some more time which is required for this.

Change-Id: I325848feb7bba25bb5dc64cea7db1dbe56c711cc
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-10172
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Shubhangi Shrivastava <shubhangi.shrivastava@intel.com>