REVERTME [VPG]: drm/i915: Fix m2 programming for CHT
PLL values to be used in CHT depends on the various calculated values
such as P1, P2, M1, M2, M2 fractional part and N. M2 out of these
should be programmed from the top byte of calculated value. Instead of
taking the topmost byte we are taking from BIT22 which results in
incorrect value programmed for M2.

This patch fixes this by correcting the shift to take the top byte
from calculated M2 for DP on CHT alone since making this for HDMI
results in regression.

REVERTME: This needs further debugging to enable the same change
for HDMI too.

v2: limit the change only for DP

Change-Id: I06406f0e573bba9dff1db185c552f812ee17623d
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5971
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/426411