EM/bq24192_charger: WA to fix the Status reg updation during reboot
In BQ CHG probe, DPDM bit is force set to true inorder to force a
DPDM detection and dectect the charger cable if it was inserted
prior to boot up.
However, when we force trigger the DPDM detection from probe,
as per the Vendor data sheet, it requires a delay of 50ms
to complete the DPDM detection.

This patch adds a flag 'dpdm_trig' which will be set during
probe when DPDM detection is set. In the IRQ thread function,
a sleep of 50ms is added if the dpdm_trig flag is set. The
flag will be susequently reset to false state.

This patch has been ported from L-Mr1-CHT_Stable branch.

Change-Id: I9eff42dfdfa7bd7e59fb9b36b6c2b5ef7abd3937
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1175
Signed-off-by: raosrini <srinidhi.rao@intel.com>
Reviewed-on: https://android.intel.com:443/425507