typec-pd: optimize the power supply calls
when the sink port negotiates for the power due to sudden
increase / decrease of power, charger has to be disconnected
and connected back. Use the power_supply notifier to
notify the change, which helps to off-load the tasks to the
power supply framework. Sink port shall update the negotiated
current using the updated interface.

Change-Id: I839598de8c6fb1e5ab21404bca42bfcf1f08889f
Signed-off-by: R, Kannappan <r.kannappan@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6023
Reviewed-on: https://android.intel.com:443/419733
Reviewed-on: https://android.intel.com:443/426208