SQUASHME! [VPG]: drm/i915: HW/SW inconsistency recovery through forced CSB check
1. This replaces forced context resubmission in the TDR/execlists API as a
method for overcoming inconsistent HW/driver states, specifically when it comes
to the observed state of what context is currently being executed
(EXECLIST_STATUS_CONTEXT_ID==X, head context of the respective execlist
queue==Y). The way it works is that if the periodic hang checker detects a
sustained, irrecoverable inconsistent HW/driver state it tries faking a context
event interrupt (since it's assumed that the state came about due to a lost
interrupt). This does not always guarantee that the inconsistency is rectified
but it typically solves the kind of HW/driver inconsistencies that happen due
to lost context completion interrupts that happen sometimes during
long-duration operations, which makes the driver think that a context is
executing on hardware when the hardware is in fact idle because it completed
the context execution without notifying the driver via the normal context
completion event interrupt.

If the faked interrupt is ineffective and the inconsistency remains this commit
introduces a promotion path from the consistency checker to full GPU reset that
allows the hang checker to skip per-engine recovery (which is ineffective
during context submission status inconsistencies) and go straight for a full
GPU reset, which is guaranteed to restore consistency by discarding all
submitted work in the driver and forcing GPU to idle and thereby returning to a
well-defined, consistent state.

    NOTICE: Promotion to full GPU reset has been disabled by default for now
    due to false positive inconsistency detections leading to unexpected full
    GPU resets.  The path is still part of the patch but the promotion decision
    is controlled through a module parameter called enable_inconsistency_reset.
    When enough testing has been done and we are 100% confident that we can
    trust the inconsistency detection we may opt to enable this feature by
    default.

In order to make full GPU reset work with not an engine hang but an
inconsistent HW/SW state which might be blocking the driver from progressing
this requires native sync to be extended with a function for resetting the
fence of a specific request that can be used for for resetting all fences for
all currently pending requests during full GPU reset in order to avoid Android
userland from timing out pending fences. This function is then integrated into
the scheduler from the point where all outstanding work is purged during a full
GPU reset. In order to make this work from the scheduler we've had to remove
the scheduler spinlock acquisition from i915_scheduler_wait_fence_signaled()
since fences, upon being signalled, sometimes call back into this function,
which results in spinlock recursion since the spinlock is already held when
signalling the fence. Not grabbing the spinlock and updating the node flag from
this function should be ok since the state of the flag is only important up
until the point where the fence is signalled and not afterwards.

2. Added a seqno hang check for the consistency checker to avoid the case where
the hang checker detects a hang because the sampled HEAD register value has
gone to zero because the hardware is idle. This is possible if two subsequent
hang check periods detect HEAD==zero because one workload goes from busy->idle
and samples HEAD==zero and then the subsequent workload goes from
idle->busy->idle and samples another HEAD==zero upon which the hang checker
decides that a hang is in progress because of these two subsequent HEAD==zero
hang sample result (no HEAD progression even though HEAD==zero is a special
case indicating hardware idle, which unfortunately the hang checker does not
understand) and calls i915_hangcheck_hung(), which then does the consistency
check and potentially finds a transient inconsistency and goes for
inconsistency rectification, which in the worst case means a full GPU reset. By
letting the consistency checker check for seqno progression two different
workloads that both result in HEAD==zero can be distinguished and even though
the hang checker sees an early hang indication the consistency checker can see
that the two HEAD==zero indications are from two different workloads and can
hold off the consistency check and thereby reduce the risk of false positive
inconsistency rectifications.

We could add the seqno check to the main hang detection algorithm in
i915_hangcheck_sample() but that sometimes makes the hang checker less
responsive, which could potentially lead to Android userland timeouts and
broken display flips etc.

3. Disabled INSTDONE checking in the hang detection algorithm since it's
unpredictable and delays hang recovery sometimes, which potentially leads to
Android userland breaking - resulting in fence timeouts and frozen display.

4. This commit also introduces a new debugfs entry point for simulating lost
context event interrupts (currently only for events that will lead to stuck
contexts in the execlist queues, since that is what the consistency checker
tries to deal with). There is a need for extended validation to test the
consistency checker since it has been shown to cause problems in the past. This
debugfs entrypoint can be used for this purpose.

	SQUASHME! - This patch should be squashed into the following patch:
		Author: Tomas Elf <tomas.elf@intel.com>
   		Change-Id: I75b1b0419a03acbd71134cfe22bf1d63ffdb0154
		REVERTME [VPG]: drm/i915: Lock-up workaround: Forced context resubmission

Change-Id: I54ce0b2da55f7229b00733ce0bd8070cc0f475dd
Signed-off-by: Tomas Elf <tomas.elf@intel.com>
Signed-off-by: Deepak S <deepak.s@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5527
Reviewed-on: https://android.intel.com:443/429007