REVERTME [VPG]: drm/i915: Re-enable primary if flip done is set on primary plane
Flip done on sprite plane will disable primary plane. Disable sprite plane is
not the only means of enabling back primary, even if flip done is set back on
primary plane, it should be enabled.

REVERTME: CHT product kernel relies on internal implementation of
atomic flip. Atomic flip is already addressed in a clean fashion using
nuclear flip in upstream and is used for Broxton and beyond.
So this patch will be a REVERTME patch and can be discarded in gmin_dev &
broxton trees and future forklifts.

Change-Id: I8057a1dc35b043ae6550b9ba8e6b443b92c4ce92
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5799
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Gajanan Bhat <gajanan.bhat@intel.com>
Reviewed-on: https://android.intel.com:443/425212