REVERTME [VPG]: drm/i915: Disable maxfifo if zorder needs to be updated
In Cherryview if display is in maxfifo mode, zorder update does not take effect.
We need to exit from maxfifo mode if enabled before updating zoder. This patch
checks if in set display function zorder update was requested and disables
maxfifo if needed.

REVERTME: CHT product kernel relies on internal implementation of
atomic flip. Atomic flip is already addressed in a clean fashion using
nuclear flip in upstream and is used for Broxton and beyond.
So this patch will be a REVERTME patch and can be discarded in gmin_dev &
broxton trees and future forklifts.

Change-Id: I1eccd6dfefd9aea4a703785aac36a2fb190b63ee
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5799
Signed-off-by: Gajanan Bhat <gajanan.bhat@intel.com>
Reviewed-on: https://android.intel.com:443/425213