REVERTME [VPG]: drm/i915: In CMD mode, Cursor Plane always enable
In Command mode, the shift screen issue observed with multi plane
enable/disable flip request. As per HW, one plane always need to be
enabled in command mode. For this the cursor plane of size 64*64
is always enable in command mode.
This is REVERTME patch, as the scenarios where cursor plane request come
from HWC are not addressed.

Change-Id: I993c0ea822fa25e2a4a08aed1b84c5607ead4ad3
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7482
Signed-off-by: Megha Aggarwal <megha.aggarwal@intel.com>
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
Reviewed-on: https://android.intel.com:443/432459