dc_xpwr_charger: Revert "REVERTME: dc_xpwr_charger: fix i2c hang issue"
This reverts commit fccb66cf3bffc6cdca9da86671cc5d0e2adfda09.

A delay of around 10ms was added to get_prpert and set_property
functions to prevent the i2c semaphore hand issue during back
to back XPWR PMIC I2 read and write calls. This fix was a
workaround till the actual fix was delivered by the BIOS team.

Since the correct fix has already been merged in BIOS, this WA
is not required any more. Hence reverting this patch.

Change-Id: I8fec7075aca54343be9c5d27ded26450e321f687
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6440
Signed-off-by: raosrini <srinidhi.rao@intel.com>
Reviewed-on: https://android.intel.com:443/428286