FOR_UPSTREAM [VPG]: drm/i915: delay short pulse handling till display is up
DPCD read will succeed in CHT only if the respective PLL & DPIO
are programmed properly, which is done as part of modeset.
Due to presence of upfront link training it is possible for sink to
give short pulse even before modeset is complete. This results in
failing to handle this short pulse. This patch delays dpcd
operations inside short pulse for a max of 80ms, which will
allow us to handle most of the short pulse generated while
modeset is in progress.

Max delay is set to 80ms instead of 100ms to account for
delays required during dpcd read.

Change-Id: Id67e8ae7701c2eed6d04b0fb715a764eac6b9b2f
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9749
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/439696