FOR_UPSTREAM [VPG]: drm/i915: Avoid enabling dithering for 6bpc
External panels does not require dithering but was still enabled
to keep it consistent with local panels. But enabling dithering
for 6bpc on external DP will result in modifying the expected
output CRC during compliance test 4.4.1.1 to 4.4.1.3 resulting
in failing those tests. This patch fixes this by avoiding
enabling of dithering for 6bpc for external display port
panels alone.

Change-Id: Id9123bb70dde3165355a9338f5744a36bef9c1f0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-27481
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/496002