Disable V1P2A rail conditionally
VIP2A rail is dedicated on FFD designs for CHT-HR/T4 but is shared
with multiple consumers (e.g. SoC I/Os, UFS, Modem, eMMC connector)
on RVP designs with WC PMIC. So this rail needs to be controlled for
FFD designs but needs to be always enabled for RVP. Therefore added a
check for the same.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9676
Change-Id: I6d8e3d63590a568fab13594e51d94a91b1507bd9
Signed-off-by: Rakesh Ranjan Jha <rakesh.r.jha@intel.com>
Reviewed-on: https://android.intel.com:443/437360