[EM/PMIC_CCSM]: Fix delay in processing cable event connect/disconnect
When the mulitple connect/disconnect events processing as the events
are getting queued, it takes lot of time to process the cable events.
So to avoid the repitation of the events, if all the cables are
disconnected remove all the previous events from the list and handle
(disconnect) only the previously connected calbes.

Change-Id: Ibdb787a8c51073050c81bb5186e7ec5dbc4b6757
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-8060
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Reviewed-on: https://android.intel.com:443/440780
Reviewed-on: https://android.intel.com:443/448858