[EM/PMIC_CCSM]: Notify OTG for all PC ports source connect/disconnect
When USB is disconnect while checking the cable type for sending
notification to OTG, if the cabe type is none not sending notification
to the otg driver. This breaks the next detections. Hence enabled the
flag to send the notification to otg for all type of pc port's connect
or disconnect only.

Change-Id: Iec557b7d7bc282443200b5460a648cbf06d8510d
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-12471
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Reviewed-on: https://android.intel.com:443/443774
Reviewed-on: https://android.intel.com:443/447383