MUST_REBASE [VPG]: drm/i915: add generic netlink msg gpu resets
Due to changes in the SELinux policy on Android, userspace is no
longer permitted to listen to uevent sockets. This causes a problem
for the OGL driver which wants to be informed of gpu resets in order
to implement the Robustness extensions. Use of the
i195_get_reset_stats_ioctl was found to incur excessive overhead
so I have added a generic netlink socket which a userland process
can listen to muticast messages from.

This is a must rebase patch because currently there is not
open source user for this. Mesa uses the ioctl. I'm not sure if
they just live with the overhead or mitigate it in some way.

The netlink message used to signal a gpu reset is just an
empty message, it contains no further information. If the
user needs more information about the reset it must use the
ioctl. Since GPU resets are rare the ioctl overhead is not
a problem if it is only called once a reset is detected.

Change-Id: I864956b900c65bd620e146bea50c844594bc170a
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-12355
Signed-off-by: Tim Gore <tim.gore@intel.com>
Reviewed-on: https://android.intel.com:443/467151