SQUASHME! [VPG]: drm/i915: Modify audio_port to identify DP/HDMI
The audio_port variable is currently used to track the port
where Audio is enabled. But this does not differentiate
HDMI vs. DP. This patch modifies the usage of 'audio_port'
as a 'intel_dig_port' pointer so that it can identify the
right encoder (HDMI/DP) connected at any point in time.

SQUASHME! - This patch should be squashed into the following existing patch:
        Author: Shubhangi Shrivastava <shubhangi.shrivastava@intel.com>
        Date: Thu, 10 Sep 2015 12:41:11 +0530
        Change-Id: I325848feb7bba25bb5dc64cea7db1dbe56c711cc
	REVERTME: [VPG]: drm/i915: Notifying audio driver on hotplug/unplug

Change-Id: I82aefcca9f39b8bcb19f569dd0110a3308aec6e6
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-10613
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Thulasimani,Sivakumar <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/448683