[EM/FUSB300]: Fix for cable detection issue
When software DRP toggle mode is enabled, both CC's are pulled up and
while measuring the voltage on the CC, individual CC's are pulled up
and DAC measurement is done. Since both CC's are pulled up the
measurement value is getting corrupted which is leading to a wrong
cable detection. Fixed by removing the pullup on both CC's as the
measurement function enables individual pullup's.

Change-Id: Ia36a07bfc519dd9806145ca9fdc60df79c88cde2
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-10109
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Signed-off-by: R, Kannappan <r.kannappan@intel.com>
Reviewed-on: https://android.intel.com:443/450975